--- packages/app/src/scenes/odysseyCreator/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './skybox.utils';
